Memory blocks can be up to 64kb in size, multiple size parameters are resized accordingly.

+++ Driver/Font/TrueType/FreeType/ttfile.c
@@ -279,7 +279,7 @@
  ******************************************************************/
 
   EXPORT_FUNC
-  TT_Error  TT_Access_Frame( STREAM_ARGS FRAME_ARGS Long  size )
+  TT_Error  TT_Access_Frame( STREAM_ARGS FRAME_ARGS Short  size )
   {
     TT_Error  error;
 
@@ -346,7 +346,7 @@
  ******************************************************************/
 
   EXPORT_FUNC
-  TT_Error  TT_Check_And_Access_Frame( STREAM_ARGS FRAME_ARGS Long  size )
+  TT_Error  TT_Check_And_Access_Frame( STREAM_ARGS FRAME_ARGS Short  size )
   {
     TT_Error  error;
     Long      readBytes, requested;
@@ -528,7 +528,7 @@
  ******************************************************************/
 
   EXPORT_FUNC
-  TT_Error  TT_Access_Frame( STREAM_ARGS FRAME_ARGS Long  size )
+  TT_Error  TT_Access_Frame( STREAM_ARGS FRAME_ARGS Short  size )
   {
     TT_Error  error;
 
@@ -584,7 +584,7 @@
  ******************************************************************/
 
   EXPORT_FUNC
-  TT_Error  TT_Check_And_Access_Frame( STREAM_ARGS FRAME_ARGS Long  size )
+  TT_Error  TT_Check_And_Access_Frame( STREAM_ARGS FRAME_ARGS Short  size )
   {
     TT_Error  error;
     Long      readBytes;
@@ -911,9 +911,9 @@
  ******************************************************************/
 
   EXPORT_FUNC
-  TT_Error  TT_Read_File( STREAM_ARGS void*  buffer, Long  count )
+  TT_Error  TT_Read_File( STREAM_ARGS void*  buffer, Short  count )
   {
-    if ( FileRead( CUR_Stream->file, buffer, count, FALSE ) != (ULong)count )
+    if ( FileRead( CUR_Stream->file, buffer, count, FALSE ) != count )
       return TT_Err_Invalid_File_Read;
 
     return TT_Err_Ok;
